Consolidate member related admin endpoints
